docs: refactor "structure" example to be exemplary for v1
- split demo code from contract
